# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vesta wardrobe app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClothingItem {
  id: number;
  userId: string;
  name: string;
  category: string;
  color: string;
  pattern: string;
  fit: string;
  style: string;
  season: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateClothingItemBody {
  name: string;
  category: string;
  color: string;
  pattern: string;
  fit: string;
  style: string;
  season: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface UpdateClothingItemBody {
  name?: string;
  category?: string;
  color?: string;
  pattern?: string;
  fit?: string;
  style?: string;
  season?: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ToggleAvailabilityBody {
  isAvailable: boolean;
}

export interface MatchRequestBody {
  clothingItemId: number;
  /** @nullable */
  filterStyle?: string | null;
}

export interface OutfitMatch {
  item: ClothingItem;
  score: number;
  reason: string;
  matchType: string;
}

export interface GenerateOutfitBody {
  /** @nullable */
  filterStyle?: string | null;
}

export interface GeneratedOutfit {
  top: ClothingItem;
  bottom: ClothingItem;
  score: number;
  reason: string;
}

export interface SaveFavoriteBody {
  topId: number;
  bottomId: number;
  /** @nullable */
  name?: string | null;
  score: number;
  reason: string;
}

export interface FavoriteOutfit {
  id: number;
  userId: string;
  topId: number;
  bottomId: number;
  /** @nullable */
  name?: string | null;
  score: number;
  reason: string;
  top: ClothingItem;
  bottom: ClothingItem;
  createdAt: string;
}

export type DashboardSummaryByCategory = { [key: string]: number };

export type DashboardSummaryByStyle = { [key: string]: number };

export type DashboardSummaryBySeason = { [key: string]: number };

export interface DashboardSummary {
  totalItems: number;
  totalFavorites: number;
  byCategory: DashboardSummaryByCategory;
  byStyle: DashboardSummaryByStyle;
  bySeason: DashboardSummaryBySeason;
}

export type ListClothingItemsParams = {
  category?: string;
  style?: string;
  season?: string;
  color?: string;
  search?: string;
  isAvailable?: boolean;
};