# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vesta wardrobe app API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: clothing
    description: Clothing item operations
  - name: outfits
    description: Outfit operations
  - name: favorites
    description: Favorite outfit operations
  - name: dashboard
    description: Dashboard summary operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /clothing:
    get:
      operationId: listClothingItems
      tags: [clothing]
      summary: List all clothing items for the authenticated user
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: style
          in: query
          schema:
            type: string
        - name: season
          in: query
          schema:
            type: string
        - name: color
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: isAvailable
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of clothing items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClothingItem"
    post:
      operationId: createClothingItem
      tags: [clothing]
      summary: Create a new clothing item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClothingItemBody"
      responses:
        "201":
          description: Created clothing item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClothingItem"

  /clothing/{id}:
    get:
      operationId: getClothingItem
      tags: [clothing]
      summary: Get a single clothing item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Clothing item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClothingItem"
        "404":
          description: Not found
    patch:
      operationId: updateClothingItem
      tags: [clothing]
      summary: Update a clothing item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClothingItemBody"
      responses:
        "200":
          description: Updated clothing item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClothingItem"
        "404":
          description: Not found
    delete:
      operationId: deleteClothingItem
      tags: [clothing]
      summary: Delete a clothing item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /clothing/{id}/availability:
    patch:
      operationId: toggleClothingAvailability
      tags: [clothing]
      summary: Toggle the availability of a clothing item (laundry mode)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleAvailabilityBody"
      responses:
        "200":
          description: Updated clothing item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClothingItem"

  /outfits/match:
    post:
      operationId: getOutfitMatches
      tags: [outfits]
      summary: Get outfit match recommendations for a clothing item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequestBody"
      responses:
        "200":
          description: List of outfit matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OutfitMatch"

  /outfits/generate:
    post:
      operationId: generateOutfit
      tags: [outfits]
      summary: Generate the best full outfit combination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOutfitBody"
      responses:
        "200":
          description: Generated outfit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedOutfit"

  /favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      summary: List all saved favorite outfits
      responses:
        "200":
          description: List of favorites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FavoriteOutfit"
    post:
      operationId: saveFavorite
      tags: [favorites]
      summary: Save a favorite outfit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveFavoriteBody"
      responses:
        "201":
          description: Saved favorite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteOutfit"

  /favorites/{id}:
    delete:
      operationId: deleteFavorite
      tags: [favorites]
      summary: Delete a favorite outfit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/recent:
    get:
      operationId: getRecentItems
      tags: [dashboard]
      summary: Get recently added clothing items
      responses:
        "200":
          description: Recent items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClothingItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ClothingItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        category:
          type: string
        color:
          type: string
        pattern:
          type: string
        fit:
          type: string
        style:
          type: string
        season:
          type: string
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - name
        - category
        - color
        - pattern
        - fit
        - style
        - season
        - isAvailable
        - createdAt
        - updatedAt

    CreateClothingItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        color:
          type: string
        pattern:
          type: string
        fit:
          type: string
        style:
          type: string
        season:
          type: string
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required:
        - name
        - category
        - color
        - pattern
        - fit
        - style
        - season

    UpdateClothingItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        color:
          type: string
        pattern:
          type: string
        fit:
          type: string
        style:
          type: string
        season:
          type: string
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]

    ToggleAvailabilityBody:
      type: object
      properties:
        isAvailable:
          type: boolean
      required:
        - isAvailable

    MatchRequestBody:
      type: object
      properties:
        clothingItemId:
          type: integer
        filterStyle:
          type: ["string", "null"]
      required:
        - clothingItemId

    OutfitMatch:
      type: object
      properties:
        item:
          $ref: "#/components/schemas/ClothingItem"
        score:
          type: number
        reason:
          type: string
        matchType:
          type: string
      required:
        - item
        - score
        - reason
        - matchType

    GenerateOutfitBody:
      type: object
      properties:
        filterStyle:
          type: ["string", "null"]

    GeneratedOutfit:
      type: object
      properties:
        top:
          $ref: "#/components/schemas/ClothingItem"
        bottom:
          $ref: "#/components/schemas/ClothingItem"
        score:
          type: number
        reason:
          type: string
      required:
        - top
        - bottom
        - score
        - reason

    SaveFavoriteBody:
      type: object
      properties:
        topId:
          type: integer
        bottomId:
          type: integer
        name:
          type: ["string", "null"]
        score:
          type: number
        reason:
          type: string
      required:
        - topId
        - bottomId
        - score
        - reason

    FavoriteOutfit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        topId:
          type: integer
        bottomId:
          type: integer
        name:
          type: ["string", "null"]
        score:
          type: number
        reason:
          type: string
        top:
          $ref: "#/components/schemas/ClothingItem"
        bottom:
          $ref: "#/components/schemas/ClothingItem"
        createdAt:
          type: string
      required:
        - id
        - userId
        - topId
        - bottomId
        - score
        - reason
        - top
        - bottom
        - createdAt

    DashboardSummary:
      type: object
      properties:
        totalItems:
          type: integer
        totalFavorites:
          type: integer
        byCategory:
          type: object
          additionalProperties:
            type: integer
        byStyle:
          type: object
          additionalProperties:
            type: integer
        bySeason:
          type: object
          additionalProperties:
            type: integer
      required:
        - totalItems
        - totalFavorites
        - byCategory
        - byStyle
        - bySeason