# lib/api-zod/src/generated/types/clothingItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vesta wardrobe app API
 * OpenAPI spec version: 0.1.0
 */

export interface ClothingItem {
  id: number;
  userId: string;
  name: string;
  category: string;
  color: string;
  pattern: string;
  fit: string;
  style: string;
  season: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable: boolean;
  createdAt: string;
  updatedAt: string;
}