# lib/api-zod/src/generated/types/createClothingItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vesta wardrobe app API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClothingItemBody {
  name: string;
  category: string;
  color: string;
  pattern: string;
  fit: string;
  style: string;
  season: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}