# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vesta wardrobe app API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryByCategory } from "./dashboardSummaryByCategory";
import type { DashboardSummaryBySeason } from "./dashboardSummaryBySeason";
import type { DashboardSummaryByStyle } from "./dashboardSummaryByStyle";

export interface DashboardSummary {
  totalItems: number;
  totalFavorites: number;
  byCategory: DashboardSummaryByCategory;
  byStyle: DashboardSummaryByStyle;
  bySeason: DashboardSummaryBySeason;
}