# lib/api-zod/src/generated/types/favoriteOutfit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vesta wardrobe app API
 * OpenAPI spec version: 0.1.0
 */
import type { ClothingItem } from "./clothingItem";

export interface FavoriteOutfit {
  id: number;
  userId: string;
  topId: number;
  bottomId: number;
  /** @nullable */
  name?: string | null;
  score: number;
  reason: string;
  top: ClothingItem;
  bottom: ClothingItem;
  createdAt: string;
}