# lib/api-zod/src/generated/types/generatedOutfit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vesta wardrobe app API
 * OpenAPI spec version: 0.1.0
 */
import type { ClothingItem } from "./clothingItem";

export interface GeneratedOutfit {
  top: ClothingItem;
  bottom: ClothingItem;
  score: number;
  reason: string;
}