# lib/api-zod/src/generated/types/matchRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vesta wardrobe app API
 * OpenAPI spec version: 0.1.0
 */

export interface MatchRequestBody {
  clothingItemId: number;
  /** @nullable */
  filterStyle?: string | null;
}